requirements.txt
numpy
matplotlib
pyserial
pandas
requests
scipy
pydantic
scipy
pydantic
pyyaml
PyQt6
pyqtgraph
xarm
hidapi
opencv-python
godirect
bleak
Flask
